after increasing size of pattern from 32 to 64 bits (COMP-2146) resolution also became 64 bits
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@94986 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/jbaci/src/alma/ACS/impl/TypelessPropertyImpl.java
@@ -51,7 +51,7 @@ public class TypelessPropertyImpl
 	/**
 	 * Property resolution (bitmask).
 	 */
-	protected int resolution;
+	protected long resolution;
 
 	/**
 	 * Constructor.
@@ -83,7 +83,7 @@ public class TypelessPropertyImpl
 			characteristic = "units";
 			units = characteristicModelImpl.getString(characteristic);
 			characteristic = "resolution";
-			resolution = characteristicModelImpl.getInteger(characteristic);
+			resolution = characteristicModelImpl.getLong(characteristic);
 		}
 		catch (Throwable t)
 		{
@@ -118,7 +118,7 @@ public class TypelessPropertyImpl
 	/**
 	 * @see alma.ACS.TypelessPropertyOperations#resolution()
 	 */
-	public int resolution() {
+	public long resolution() {
 		return resolution;
 	}
